+++ src/mbgl/util/mat4.cpp
@@ -46,7 +46,7 @@ void matrix::identity(mat4& out) {
 }
 
 bool matrix::invert(mat4& out, mat4& a) {
-    float a00 = a[0], a01 = a[1], a02 = a[2], a03 = a[3],
+    double a00 = a[0], a01 = a[1], a02 = a[2], a03 = a[3],
           a10 = a[4], a11 = a[5], a12 = a[6], a13 = a[7],
           a20 = a[8], a21 = a[9], a22 = a[10], a23 = a[11],
           a30 = a[12], a31 = a[13], a32 = a[14], a33 = a[15],
